Use `process_runner` for Sharezone CLI. (#1131)
By using [`process_runner`](https://pub.dev/packages/process_runner)
(which uses the [`process`](https://pub.dev/packages/process) package)
we:

**Make testing our CLI easier in the future**.   
`process_runner` uses the `ProcessManager` class from the `process`
package internally, which is used to provide a level of indirection from
the static methods in the
[Process](https://api.dart.dev/stable/3.1.3/dart-io/Process-class.html)
class, which in turn allows the underlying implementation to be mocked
out or decorated for testing and debugging purposes.

**Remove custom hacks**  
In my custom `runProcess` method I had to work around some quirks of the
Dart native Process handling. `process_runner` deals with that as well,
so that we have less custom code to worry about.