Upgrade to Flutter v3.16.3 (#1164)
## Description

This PR upgrades to Flutter v3.16.3. We faced the following breaking
changes:

* [Deprecated just-in-time navigation pop APIs for Android Predictive
Back](https://docs.flutter.dev/release/breaking-changes/android-predictive-back)
* [The ThemeData.useMaterial3 property is now set to true by
default](https://docs.flutter.dev/release/breaking-changes/material-3-default)

## Material 3

For now, we just opt out of Material 3 because it requires some more
manual migration and isn't our current focus (#1159).

## Predictive Back

> Android 14 introduced the [Predictive Back
feature](https://developer.android.com/guide/navigation/predictive-back-gesture),
which allows the user to peek behind the current route during a valid
back gesture and decide whether to continue back or to cancel the
gesture. This was incompatible with Flutter’s navigation APIs that allow
the developer to cancel a back gesture after it is received.
>
> With predictive back, the back animation begins immediately when the
user initiates the gesture and before it has been committed. There is no
opportunity for the Flutter app to decide whether it’s allowed to happen
at that time. It must be known ahead of time.
>
> For this reason, all APIs that allow a Flutter app developer to cancel
a back navigation at the time that a back gesture is received are now
deprecated. They have been replaced with equivalent APIs that maintain a
boolean state at all times that dictates whether or not back navigation
is possible. When it is, the predictive back animation happens as usual.
Otherwise, navigation is stopped. In both cases, the app developer is
informed that a back was attempted and whether it was successful.

From
https://docs.flutter.dev/release/breaking-changes/android-predictive-back#background

We mostly have used `WillPopScope` for confirmation dialogs, so the
migration looked like this:

```dart
WillPopScope(
  onWillPop: () async {
    final bool? shouldPop = await _showBackDialog();
    return shouldPop ?? false;
  },
  child: child,
)
```

to

```dart
return PopScope(
  canPop: false,
  onPopInvoked: (bool didPop) async {
    if (didPop) {
      return;
    }
    final NavigatorState navigator = Navigator.of(context);
    final bool? shouldPop = await _showBackDialog();
    if (shouldPop ?? false) {
      navigator.pop();
    }
  },
  child: child,
)
```


https://docs.flutter.dev/release/breaking-changes/android-predictive-back#migrating-a-back-confirmation-dialog

### Run the application with predictive back

1. Run Android 33 or above.
2. Enable the feature flag for predictive back on the device under
“Developer options”. This will be unnecessary on future versions of
Android.
3. Set `android:enableOnBackInvokedCallback="true"` in
android/app/src/main/AndroidManifest.xml. If needed, refer to [Android’s
full
guide](https://developer.android.com/guide/navigation/custom-back/predictive-back-gesture)
for migrating Android apps to support predictive back.
4. Make sure you’re using version 3.14.0-7.0.pre of Flutter or greater.
5. Run the app and perform a back gesture (swipe from the left side of
the screen).


https://docs.flutter.dev/release/breaking-changes/android-predictive-back#supporting-predictive-back

Unfortunately, I wasn't able to get predictive back animations running
in a Flutter app. They work for me in the native settings app, but not
in a Flutter app (even with the example app). I think it shouldn't be
that a big deal because it should work that way as we have it. Here is a
demo of how it should look:
https://github.com/flutter/flutter/pull/120385.

## Testing

I tested the changes manually.

---------

Co-authored-by: Jonas Sander <29028262+Jonas-Sander@users.noreply.github.com>