Move `Dimensions` to `sharezone_widgets` (#1193)
I moved the `dimensions.dart` class to the `sharezone_widgets` package.
Technically, it's not a widget but it's something that is only used in
our UI so I think it's okay to have in ``sharezone_widgets` (better
ideas are welcome). This removes the direct dependency of
`sharezone_utils` in `sharezone_widgets`.