Use Flutter/Dart SDK in Sharezone Repo CLI, if FVM is not installed (#1196)
I tried to fix #1195 and thought maybe it's related to the FVM cache.
Therefore, I added a `runDartCommand` and `runFlutterCommand` that uses
the FVM if it's installed. In case, it's not installed, it uses the
Flutter/Dart SDK. However, it haven't fixed the issue but I think it's
still helpful because it simplifies the GitHub Actions.