Convert `lightTheme` and `darkTheme` to methods (#1234)
This PR converts the variables `lightTheme` and `darkTheme` to the
methods `getLightTheme()` and `getDarkTheme()`. The reason for this is
that in golden tests outside of `/app`, the `golden_toolkit` package is
unable to load our "Rubik" font (see
https://github.com/eBay/flutter_glove_box/issues/158). With this method,
the golden tests can override the font family with a font like "Roboto".