Remove `timestampFromDateTime` method (#1235)
It's not necessary to convert a `DateTime` to a `Timestamp`. The
`cloud_firestore` package already does this for you. The background for
removing this method is because I want to get rid of the
`cloud_firestore` dependency in `firebase_helper.dart` to avoid having
the Firebase dependencies on our website.