Move widgets of Sharezone Plus page into it's own package (#1233)
This PR moves the UI widgets of the Sharezone Plus page into its own
package called `sharezone_plus_page_ui`. The motivation behind this is
to reuse these widgets on our website. This package only contains the UI
not any logic, like `SharezonePlusPageController` because the logic is
different between our app and the website (in our app you have an
account, on our website you don't have an account).