Add prototype UI for a single-page dialog for adding events/exams. (#1286)
This PR adds a a prototype UI for adding events/exams. Pressing buttons
etc doesn't do anything yet.

I copied most of the code it from `homework_dialog.dart`. In the next PR
I want to move shared widgets out to `sharezone_widgets`. I commented
some stuff out so I don't forget some functionality in the future.

Since the new dialog doesn't work yet and we can't see in the code if
we're in a preview, I disabled showing this dialog for now.

|
![grafik](https://github.com/SharezoneApp/sharezone-app/assets/29028262/074ca097-5886-4216-b2f9-76ec5f6fecf5)
|
![grafik](https://github.com/SharezoneApp/sharezone-app/assets/29028262/498f1b5f-19bb-4478-a6ce-718f04dae975)
|
|---|---|


![grafik](https://github.com/SharezoneApp/sharezone-app/assets/29028262/2f897c06-a419-4525-a8c9-b3039d39ff1b)


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced the timetable functionality with the ability to add events
through a dialog or a dedicated page, depending on specific conditions.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->