Replace `AlphaVersionBanner` with `DevelopmentStageBanner`. (#1291)
This replaces the `AlphaVersionBanner` with `DevelopmentStageBanner` so
that a banner is always shown if the user uses a non-stable version of
Sharezone.
This is useful for #1290 (new "preview" stage for preview builds in
PRs).

Currently the banner is only shown in alpha, not in beta. This will show
a banner in alpha and beta.


![grafik](https://github.com/SharezoneApp/sharezone-app/assets/29028262/44bcbfa5-8137-4fe2-81f0-a0cc7db2e9d1)

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Introduced a Development Stage Banner to inform users about the
current development stage of the app.
- **Refactor**
- Improved the flexibility of testing by allowing the development stage
variable to be overridden.
- Enhanced the Development Stage Banner to dynamically display the app's
development stage.
- **Tests**
- Added tests to ensure the Development Stage Banner accurately reflects
the app's current stage.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->