Randomize tests; add `--test-randomize-ordering-seed` option to `sz test`. (#1318)
By default Dart/Flutter always run tests in the same order. This can be
changed by specifying `--test-randomize-ordering-seed` flag. We added
this flag to `sz test` and also use the value "random" by default.

If a test fails because of the randomization one can see that seed at
the very start of the output of the failing test. This seed can then be
passed like this: `sz test --test-randomize-ordering-seed 1784202788`.

Requires #1319 

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Added an option to specify a seed for randomizing test case execution
order.
- Users can now provide a 32-bit unsigned integer or "random" as the
seed.
- Improves test robustness by allowing control over test case execution
order.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->