Always ask the user with Google account to use (#1329)
When a user isn't sure which Google account was the right one, signs in
with the wrong account, and then wants to sign in with the correct
account, Android just used the same Google account as used before. The
only workaround is to delete the cache. This PR changes the behavior to
ask always for the Google Account.

I couldn't test it because I don't have the right signing key and
therefore can't use the Google Sign In. However, I will test it on the
alpha branch later.