Only deploy website when "website/**" has changed (#1352)
This PR adds the "path" property to the website CD pipeline.

Although our website depends on the `lib/` folder, I think it's better
to not always deploy a new version if only `lib/` changed and not
`website/` to reduce the deployments (every deployment makes the cache
invalid). It's very likely that in this case, nothing important changed
on the website. In case it was important, you can always trigger a
manual deployment.