Refactor course create page (#1393)
In this PR, I removed code duplication (removed a lot of code) between
create source without school class and with school class. Other changes:

* New: The course templates are no longer sorted by categories. When I
was in school, I remember that it always took me ~10 seconds to search
for a course in the course template page because there wasn't any
helpful structure. So I just sorted the courses by alphabet.
* New: When opening the create course page, we should get a warning that
these courses are not connected to the school class. For a better UX I
added a card to select the school class in case the user is admin in one
or more school classes.
* New: The confirmation bar for school class courses also shows the name
of the course.
* New: The confirmation snack bar for school class courses also has an
"Undo" button to delete the created course.
* New: Loading animation in the course tile when deleting a course from
the snack bar.
* New: When already created a course with the name of a template and
clicks on it, a confirmation dialog will be displayed.
* New: Add Ethik as subject to the templates.
* Fixed: When editing the course template and creating the source, the
created course wasn't marked as created in the list.
* Fixed: The padding of the loading circle wasn't correct.
* Changed: Changed the "Hinzufügen" tooltip of the edit icon button to
"Bearbeiten".



https://github.com/SharezoneApp/sharezone-app/assets/24459435/6e79713b-1d5e-4dde-90b3-0ad8f2fd4ab9

Demo Update (14.03.2024):


https://github.com/SharezoneApp/sharezone-app/assets/24459435/6de3968b-d144-4b4b-b886-49b24894fc79