Add chat feature to feedback (#1409)
This PR adds a chat feature to the Feedbacks. I created a
`feedback_shared_implementation` package which shares the implementation
of the `FeedbackDetailsPage` and a few widgets (like `FeedbackCard`).
When a user adds a comment, a document in the `Messages` sub collection
will be created. This triggers a cloud function that sends a push
notification and sets the `hasUnreadMessages` status.


https://github.com/SharezoneApp/sharezone-app/assets/24459435/8037d88d-0a73-44fb-b73b-d41e749e4041

Sorry for the big PR 🥲