Complete 'Add Grade Dialog' (#1456)
## Description

This PR completes the grade dialog to enter a new grade.

The dialog has the following "smart" features:
* When the dialog is opened, the active term is selected.
* If the user selects a grading system that is not the grading system of
the term, the Take Grade Into Account switch is disabled (and a note is
displayed).
* When changing the term, the grading system of the newly selected term
will be used (if the grade value is empty).
* When changing the grade type, the title will be set to the display
name of the grade type if the title isn't the display name of the
previous grade type.

## Demo


https://github.com/SharezoneApp/sharezone-app/assets/24459435/408dd30c-66d2-4fd5-ac54-352c4972fb89



https://github.com/SharezoneApp/sharezone-app/assets/24459435/652fc705-08b7-4db5-b72b-c06d820271d5


## Open questions

* When the user enters a grade (e.g. "1+") and changes the grade system
(e.g. to "15 - 0 Punkte"), should we clear the grade value field?

## Things for later
* Add tests (the controller has so much logic, you can easily write ~50
tests for it) -> ideally we add this before we add editing grades
* Select the date from the past exams (connect exam feature with grades
feature)