Add term settings page (#1464)
## Demo


https://github.com/SharezoneApp/sharezone-app/assets/24459435/4b785580-4ff3-469e-9a80-e0def2296863

## Description

This PR adds the settings page for terms.

For the subject weights: We merge all subjects and courses in the UI so
it's more intuitive to the user what the setting does. Otherwise, we
would only show the subjects for which the user has added a grade. If
the user sets the weight for a subject that doesn't exist, we will
automatically create one.

---------

Co-authored-by: Jonas <--help>