Fix: Deleting a term now sets `currentTerm` in Firestore doc to `null`. (#1575)
When deleting all terms `currentTerm` would still be set to the old
current term before. This is because if no current/active term was
existing we didn't add the `currentTerm` key to the data map at all.
This would leave the value unchanged. Now we explicitly set the field to
null in this case.

Fixes #1573