Add teachers to the timetable (#1629)
## Description

This pull request adds the Sharezone Plus feature to add teachers to the
timetable.

## Demo


https://github.com/SharezoneApp/sharezone-app/assets/24459435/51576d4d-c5ac-42be-bee9-04084cb228a5

Changed the lesson sheet to a dialog:


![image](https://github.com/SharezoneApp/sharezone-app/assets/24459435/a856afb7-9c76-4dd7-8ae0-f4af1466f6e3)

Reason: In the sheet you need to scroll to see the substitutions tiles
and it's not easy to adjust the size of the sheet (depending on the
screen size). A dialog is here the better option.

Users can also add the teacher when adding a lesson:


![image](https://github.com/SharezoneApp/sharezone-app/assets/24459435/cdbff9b2-2d99-4371-906a-414aba708264)

Sharezone Plus page:


![image](https://github.com/SharezoneApp/sharezone-app/assets/24459435/e737e6d5-5bbb-4638-ab2b-cd6d38d882d0)

## Related Tickets

Closes #1608
Opens #1628