Add teachers to substitution plan (#1647)
## Demo


https://github.com/SharezoneApp/sharezone-app/assets/24459435/f9b89af0-f914-4ad5-abca-ff6e1d927745

## Summary

A user can add substitution teachers. Free users can see the teachers
now. Substitutions are a list on the client now.

## Description

This PR adds the option to add substitution teachers.

I changed that now free users can also see the teachers. Reason for
this: It doesn't really makes sense if a user can see the substitution
teacher, but not the normal teacher.

Before this PR, a lesson had only one substitution at the time. Now, a
lesson can have multiple substitutions (like room substitution and
teacher substitution). This **not** a breaking change thanks to a robust
database design.

## Related Tickets

Closes #1628