Fix subtitle cards in documentation (#1744)
This pull request includes updates to the links in the documentation
files to ensure they point to the correct paths. The changes primarily
involve modifying the `href` attributes in several `Card` components.

Link updates in documentation:

*
[`docs/pages/account.mdx`](diffhunk://#diff-42dedcb5586d925a9dbe1c2e180d65695fe5bf6b95435882aa5b562e72dabfe4L8-R13):
Updated the `href` attributes in the `Card` components to remove the
`/docs/pages` prefix and point directly to the correct paths.
*
[`docs/pages/app-settings/notifications.mdx`](diffhunk://#diff-1e6fc0c32ae944383724d8c2daa335de4c96fca6e2f0c99fe013bc672b53a8dfL10-R18):
Updated the `href` attributes in the `Card` components to remove the
`/docs/pages` prefix and point directly to the correct paths.



https://github.com/user-attachments/assets/be1cd811-4e40-4c41-ae68-2f8cd6bc9d66