Add prototype of Google AdMob ads (#1760)
## Description

This pull request adds Google AdMob for displaying ads in Sharezone.
It's an experiment that we will run in the next weeks. Because we set
the
[`tagForUnderAgeOfConsent`](https://developers.google.com/admob/flutter/privacy/gdpr#tag_for_under_age_of_consent)
to `true`, we don't need to show a consent dialog because only ["limited
ads"](https://developers.google.com/admob/flutter/privacy/ad-serving-modes#limited_ads)
will be displayed.

Use the activation code "ads" to enable ads.

## Demo

| Dashboard | Timetable | Homework | Info Screen |
|--------|--------|-----|---------|
|
![Screenshot_20241012-221743](https://github.com/user-attachments/assets/280dadcf-4272-4507-b99c-fe31a7093521)
|
![Screenshot_20241012-222240](https://github.com/user-attachments/assets/744fba56-aa97-49e0-9c13-deaf5873dadf)
|
![Screenshot_20241012-222232](https://github.com/user-attachments/assets/e812351b-22f9-4f0e-b682-9625d13e3d82)
|
![Screenshot_20241013-184158](https://github.com/user-attachments/assets/6920ff44-0f2f-4b38-91f2-538625bf57b4)
|

Additionally, we show after every fifth homework check a full-screen ad:


https://github.com/user-attachments/assets/62ca4b3f-c30a-4835-b10d-678fddf48f1a