Fix race condition when joining a school class with courses (#1781)
As described in #1780, there was a bug where nothing happened after
selecting a course to join a class. I added some print statements to
understand the problem and found out that when pushing the
`_GroupJoinPage` page with `Navigator.push` multiple instances of the
`GroupJoinBloc` were created (~16 instances). This led to the problem
that when later accessing the `GroupJoinBloc` with
`BlocProvider.of(context)` a different bloc was returned than the one
used when entering the code.

Fixes #1780