Add developer error in grade system tests.
Adds `throwErrorForSubjectsWithNoGrades` to `GradesTestController.createTerm`. This is to prevent developers from being confused when they create a term with a subject that has no grades and then try to access the subject of the term. This would throw an error as the term won't have the subject added, as it has no grades for it (yet).