Update to Flutter 3.27.4 (#1815)
Update to Flutter 3.27.4

Notes:
Use of pub workspaces (should result in faster analyzer)

https://docs.flutter.dev/release/breaking-changes/wide-gamut-framework
We shouldn't use some deprecated `Color` attributes, but we use these in
database and some old code that needs to be changed first to accommodate
these changes. For now I'd just ignore this and fix this later (when
they are removed in the future we're forced to do this anyways).