Update Flutter to v.3.29.0 (#1818)
I made these changes:
* Formatted all files (a new formatting style is now used) 
* Replaced deprecated members (one to note: replace "dart:html"
(deprecated) with "web" and "http" package)

iOS minimum deployment target had to change to 13.0 because of
`purchases_flutter` (RevenueCat).

If you're interested on the changes in detail then look at the commits,
since the diff is full with formatting changes.

Otherwise not much stuff seems to have been changed