Fix CI/CD Pipelien for iOS and macOS (#1829)
## Changes:

### Upgrade to codemagic-cli-tools v0.57.3

Upgraded to the latest version.

### Upgrade to macos-15

> ITMS-90725: SDK version issue - This app was built with the iOS 17.5
SDK. Starting April 24, 2025, all iOS and iPadOS apps must be built with
the iOS 18 SDK or later, included in Xcode 16 or later, in order to be
uploaded to App Store Connect or submitted for distribution.

To fix this warning by Apple, we needed to upgrade from `macos-14` to
`macos-15`. `macos-15` used XCode 16, see
https://github.com/actions/runner-images/blob/main/images/macos/macos-15-Readme.md#xcode