Revert "Remove renaming the files locally" (#1868) and add `rename` to `downloadFileFromURL` method (#1884)
Unfortunately,
[`flutter_cache_manager`](https://pub.dev/packages/flutter_cache_manager)
does not use the Content-Disposition header as expected. Consequently,
both iOS and Android attempted to open a file with a UUID as its name
and no file extension. Reverting #1868 resolves this issue. However,
renaming in `downloadFileFromURL` breaks the file grid view (sometimes
the OS throws an error). To prevent this, I added a `rename` parameter
to `downloadFileFromURL`. Setting the rename parameter to `false` will
prevent the file from being renamed, which is necessary for the file
grid view.