Replace dynamic links with`app_links` package (#1914)
Use [`app_links`](https://pub.dev/packages/app_links) package since
dynamic links is deprecated.

As a quick fix we generate the joinLinks locally since regenerating all
links in the backend will take too long until the service is deprecated.

- [x] Android
- [x] iOS 
- [ ] macOS (should work in theory but didn't when testing locally)
- [ ] Web (?)

On macOS in the `app_links` docs it only says to add something to the
`Info.plist` file, but the AI says that it has to be added as a
"Associated Domains" capability, as the `Info.plist` is only for
schemes, not `https`-URLs. I've added the domain under "Associated
Domains" like for iOS and not added anything to the plist file.