Add clock easter egg (#1918)
Hi!

This PR is related to task #1071.

When I was going through the registration flow and saw the clock icon on
the benefits screen, a funny idea came to mind: why not make it show the
actual time?

That’s how this little easter egg was born:

Now, instead of using clock.svg, the benefits screen displays an analog
clock widget that closely resembles the original SVG image.
The clock is fully drawn with a RenderObject for better performance.
If you tap it, the hands smoothly “catch up” to the real time with an
animation. Tapping again resets them.

It’s a small detail, but I think it makes the registration process a bit
more fun and fits nicely with the theme of saving time. Hope you’ll
enjoy it!

You can also find a usage example here:
https://dartpad.dev/?id=0be0502c8caa5e73bfe650365c1084fd